In the name of innovation... * Update modules/cluster/gke to incorporate maintenance and build clusters with images from rapid release channel * Update cf4k8s experiment and modules to auto-gen certs for system and workloads domains * Bring forward ytt-libs/cf4k8s vendir support for cf-for-k8s config to commit e61ba6425502077e9daf2f78fcd2c697b76cb4e3 * Update documentation * Refine cleanup script

+++ experiments/k8s/cf4k8s/destroy-cf4k8s.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
 
+if [ -z "$1" ]; then
+	echo "Usage: create-cf4k8s.sh {iaas}"
+	exit 1
+fi
+
+IAAS="$1"
+
 terraform destroy -auto-approve
+
+rm -f ../../../modules/cf4k8s/certs.auto.tfvars
 rm -f ../../../modules/cf4k8s/templates/config.yml
+rm -Rf  ../../../modules/cf4k8s/acme/$IAAS/.terraform ../../../modules/cf4k8s/acme/$IAAS/terraform.log ../../../modules/cf4k8s/acme/$IAAS/terraform.tfstate*
 rm -Rf ../../../ytt-libs/cf4k8s/vendor
 rm -f ../../../ytt-libs/cf4k8s/vendir.lock.yml
